backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bcrypt==4.3.0",
    "fastapi>=0.119.1",
    "jwt>=1.4.0",
    "passlib>=1.7.4",
    "passlib[bcrypt]<2.0.0,>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0"
]
